Fix crash when background location permission not granted
