+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/ui/Utils.kt
@@ -1,11 +1,14 @@
 package com.websmithing.gpstracker2.ui
 
+import android.Manifest
 import android.app.Activity
 import android.content.Context
+import android.content.pm.PackageManager
 import android.location.Location
 import android.widget.Toast
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.platform.LocalContext
+import androidx.core.content.ContextCompat
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelStoreOwner
@@ -81,3 +84,10 @@ inline fun <reified VM : ViewModel> activityHiltViewModel(): VM {
 }
 
 fun hasSpaces(str: String) = str.contains(' ')
+
+fun isBackgroundLocationPermissionGranted(context: Context): Boolean {
+    return ContextCompat.checkSelfPermission(
+        context,
+        Manifest.permission.ACCESS_BACKGROUND_LOCATION
+    ) == PackageManager.PERMISSION_GRANTED
+}
